+++ .idea/libraries/Maven__com_google_protobuf_protobuf_java_3_19_4.xml
@@ -1,13 +1,13 @@
 <component name="libraryTable">
   <library name="Maven: com.google.protobuf:protobuf-java:3.19.4">
     <CLASSES>
-      <root url="jar://F:/maven/repository/com/google/protobuf/protobuf-java/3.19.4/protobuf-java-3.19.4.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/com/google/protobuf/protobuf-java/3.19.4/protobuf-java-3.19.4.jar!/" />
     </CLASSES>
     <JAVADOC>
-      <root url="jar://F:/maven/repository/com/google/protobuf/protobuf-java/3.19.4/protobuf-java-3.19.4-javadoc.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/com/google/protobuf/protobuf-java/3.19.4/protobuf-java-3.19.4-javadoc.jar!/" />
     </JAVADOC>
     <SOURCES>
-      <root url="jar://F:/maven/repository/com/google/protobuf/protobuf-java/3.19.4/protobuf-java-3.19.4-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/com/google/protobuf/protobuf-java/3.19.4/protobuf-java-3.19.4-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
