+++ .idea/libraries/Maven__junit_junit_4_13_2.xml
@@ -1,13 +1,13 @@
 <component name="libraryTable">
   <library name="Maven: junit:junit:4.13.2">
     <CLASSES>
-      <root url="jar://F:/maven/repository/junit/junit/4.13.2/junit-4.13.2.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/junit/junit/4.13.2/junit-4.13.2.jar!/" />
     </CLASSES>
     <JAVADOC>
-      <root url="jar://F:/maven/repository/junit/junit/4.13.2/junit-4.13.2-javadoc.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/junit/junit/4.13.2/junit-4.13.2-javadoc.jar!/" />
     </JAVADOC>
     <SOURCES>
-      <root url="jar://F:/maven/repository/junit/junit/4.13.2/junit-4.13.2-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/junit/junit/4.13.2/junit-4.13.2-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
