+++ .idea/libraries/Maven__mysql_mysql_connector_java_8_0_30.xml
@@ -1,13 +1,13 @@
 <component name="libraryTable">
   <library name="Maven: mysql:mysql-connector-java:8.0.30">
     <CLASSES>
-      <root url="jar://F:/maven/repository/mysql/mysql-connector-java/8.0.30/mysql-connector-java-8.0.30.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/mysql/mysql-connector-java/8.0.30/mysql-connector-java-8.0.30.jar!/" />
     </CLASSES>
     <JAVADOC>
-      <root url="jar://F:/maven/repository/mysql/mysql-connector-java/8.0.30/mysql-connector-java-8.0.30-javadoc.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/mysql/mysql-connector-java/8.0.30/mysql-connector-java-8.0.30-javadoc.jar!/" />
     </JAVADOC>
     <SOURCES>
-      <root url="jar://F:/maven/repository/mysql/mysql-connector-java/8.0.30/mysql-connector-java-8.0.30-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/mysql/mysql-connector-java/8.0.30/mysql-connector-java-8.0.30-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
