+++ .idea/libraries/Maven__org_jetbrains_annotations_23_0_0.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.jetbrains:annotations:23.0.0">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/jetbrains/annotations/23.0.0/annotations-23.0.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/jetbrains/annotations/23.0.0/annotations-23.0.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/jetbrains/annotations/23.0.0/annotations-23.0.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
