+++ .idea/libraries/Maven__org_junit_jupiter_junit_jupiter_api_5_9_1.xml
@@ -1,13 +1,13 @@
 <component name="libraryTable">
   <library name="Maven: org.junit.jupiter:junit-jupiter-api:5.9.1">
     <CLASSES>
-      <root url="jar://F:/maven/repository/org/junit/jupiter/junit-jupiter-api/5.9.1/junit-jupiter-api-5.9.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/junit/jupiter/junit-jupiter-api/5.9.1/junit-jupiter-api-5.9.1.jar!/" />
     </CLASSES>
     <JAVADOC>
-      <root url="jar://F:/maven/repository/org/junit/jupiter/junit-jupiter-api/5.9.1/junit-jupiter-api-5.9.1-javadoc.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/junit/jupiter/junit-jupiter-api/5.9.1/junit-jupiter-api-5.9.1-javadoc.jar!/" />
     </JAVADOC>
     <SOURCES>
-      <root url="jar://F:/maven/repository/org/junit/jupiter/junit-jupiter-api/5.9.1/junit-jupiter-api-5.9.1-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/junit/jupiter/junit-jupiter-api/5.9.1/junit-jupiter-api-5.9.1-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
