+++ .idea/libraries/Maven__org_junit_platform_junit_platform_commons_1_9_1.xml
@@ -1,13 +1,13 @@
 <component name="libraryTable">
   <library name="Maven: org.junit.platform:junit-platform-commons:1.9.1">
     <CLASSES>
-      <root url="jar://F:/maven/repository/org/junit/platform/junit-platform-commons/1.9.1/junit-platform-commons-1.9.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/junit/platform/junit-platform-commons/1.9.1/junit-platform-commons-1.9.1.jar!/" />
     </CLASSES>
     <JAVADOC>
-      <root url="jar://F:/maven/repository/org/junit/platform/junit-platform-commons/1.9.1/junit-platform-commons-1.9.1-javadoc.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/junit/platform/junit-platform-commons/1.9.1/junit-platform-commons-1.9.1-javadoc.jar!/" />
     </JAVADOC>
     <SOURCES>
-      <root url="jar://F:/maven/repository/org/junit/platform/junit-platform-commons/1.9.1/junit-platform-commons-1.9.1-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../../BuildTools/m2/repository/org/junit/platform/junit-platform-commons/1.9.1/junit-platform-commons-1.9.1-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
